Crate tozny_auth

Library interface to the Tozny authentication service. The purpose of this SDK is to make it easy to add Tozny support to Rust apps.

For a working example that uses this library, see tozny-pam.

Reexports

pub use self::login::{Login};

pub use self::realm::{Realm};

pub use self::user::{User, UserApi};

Modules

login

protocol
This module defines several types and provides some low-level functions.

question

realm
High level methods for making realm-level API calls to the Tozny API.

user
High level methods for making realm-level API calls to the Tozny API.