Module tozny_auth::protocol

This module defines several types and provides some low-level functions.

Structs

Challenge

KeyId

Method

Presence

Secret

SessionId

SignatureType

Timestamp
Wraps a DateTime<UTC> value; implements serialization into the format expected by the Tozny API.

UserId

Traits

Newtype
Abstraction for a type wrapper around a generic type.

Functions

error_response
Extracts error messages from a Tozny API response.