Module tozny_auth::question

Structs

Question
Type representing a signed message. The data in a Question is signed using HMAC-SHA256.

Enums

QuestionError
Enumerates the possible errors that may occur while signing a message, verifying the signature of a message, or transmitting a signed message to the Tozny API.

Functions

check_signature
Verifies a signature using a constant-time comparison.

from_json
Gets a value of a Decodable type out of a Json value.

send_request
Low-level function to dispatch a Question to the Tozny API.

sign
Produces a signature using HMAC-SHA256.

translate_url
Translates a url::Url value to a hyper::Url value.

unpack
Unpacks a base64-encoded JSON value.