Enum tozny_auth::question::QuestionError

pub enum QuestionError {
    DecoderError(DecoderError),
    EncoderError(EncoderError),
    ParserError(ParserError),
    Base64Error(FromBase64Error),
    HttpError(HttpError),
    IoError(IoError),
    Utf8Error(Utf8Error),
    InvalidSignature,
    BadlyFormedResponse,
    ErrorResponse(Json),
}

Enumerates the possible errors that may occur while signing a message, verifying the signature of a message, or transmitting a signed message to the Tozny API.

Variants

DecoderError

EncoderError

ParserError

Base64Error

HttpError

IoError

Utf8Error

InvalidSignature

BadlyFormedResponse

ErrorResponse