Module tozny_auth::user

High level methods for making realm-level API calls to the Tozny API.

API calls defined in this module do not require authentication.

Structs

LoginChallenge
Result of login_challenge call. Contains a number of values that are necessary for an authentication flow. A brief rundown:

User
Information associated with a Tozny user. This struct should be expanded in the future.

UserApi
Interface for sending user-level API calls to Tozny.