Struct tozny_auth::user::LoginChallenge

pub struct LoginChallenge {
    pub challenge: Challenge,
    pub realm_key_id: KeyId,
    pub session_id: SessionId,
    pub qr_url: Url,
    pub mobile_url: Url,
    pub created_at: Timestamp,
    pub presence: Presence,
}

Result of login_challenge call. Contains a number of values that are necessary for an authentication flow. A brief rundown:

qr_url is the URL of a QR code that can be displayed to the user. The user authenticates by scanning the code with the Tozny mobile app.

mobile_url is the URL encoded in that QR code. It uses a custom scheme. If opened on a device with the Tozny app installed, the app will open automatically.

presence this value may be stored for future logins. If a stored presence value from a previous login is available, it may be used with the push method to send a push notification to the user's mobile device instead of (or in addition to) displaying a QR code.

session_id is used with check_session_status determine whether the user has completed authentication.

Fields

challenge

realm_key_id

session_id

qr_url

mobile_url

created_at

presence